LIBASPACE-281. Modified config.rb to distinguish between server URLs
Modified "archivesspace/config/config.rb" to use different environment
variables for pointing the backend ports and by the GUI to point to the
server.

The SERVER_NAME and PUBLIC_SERVER_NAME variables were being used for
both, but in Kubernetes there is a distinction.

The SERVER_NAME and PUBLIC_SERVER_NAME are now only used internally to
point to the actual server ports on the application.

Added STAFF_INTERFACE_PROXY_URL and PUBLIC_INTERFACE_PROXY_URL to
point to the URLs that actually are displayed in the GUI.

https://issues.umd.edu/browse/LIBASPACE-281