LIBASPACE-292. Changed base image to "openjdk:8u265-jre"
Followed example from https://github.com/dartmouth-dltg/aspace-docker
and changed base image from "ubuntu:18.04" to "openjdk:8u265-jre",
which matches the version on the ArchivesSpace server.

This eliminates an error when generating PDFs, as the the Ubuntu
images places an "accessibility.properties" file into the JRE
installation which references a class which is not available. Using
the "openjdk:8u265-jre" image eliminates this problem.

https://issues.umd.edu/browse/LIBASPACE-292