LIBASPACE-289. Initial implementation of api-proxy Dockerfile and config
An "api-proxy" Dockerfile and configuration using Nginx.

The "api-proxy" container is expected to be inserted between the Nginx
running in Kubernetes and the ArchivesSpace application for requests
to the ArchivesSpace API. The proxy will then reject requests (other
than to the "login" endpoint) that do not have an
"X-ArchivesSpace-Session" header. The ArchivesSpace API will validate
the header.

This should prevent anonymous access to the ArchivesSpace API, while
allowing access to users with ArchivesSpace credentials.

https://issues.umd.edu/browse/LIBASPACE-289