LIBASPACE-289. Modified Nginx config to pass login HTTP query params
Modified the Nginx configuration for the api-proxy so that calls to
the "login" endpoint will have their query parameters sent to the
proxied server.

This is needed because the ASnakeClient
(https://github.com/archivesspace-labs/ArchivesSnake) submits a POST
request for /users/jsmith/login, but puts the parameters in the URL
instead of in the POST message body, so the submitted URL looks like
/users/jsmith/login?password=XXXXX&expiring=False, but after going
through the proxy, ArchivesSpace was just receiving
"/users/jsmith/login" and rejecting it because a password was not
provided.

While it is arguably bad form for the password to be sent in HTTP
query parameters in this manner (as the password shows up in clear
text in logs, such as the api-proxy log), this is an ArchivesSnake issue
that needs to be fixed in that library.

https://issues.umd.edu/browse/LIBASPACE-289