Revert "LIBASPACE-298. Added proxy_protocol line to api-proxy Nginx config"
This reverts commit 7ae976841688566b98025f2b405551bd9f7e8a07.

The previous commit was created during the ArchivesSpace migration to Kubernetes
and intended as a hot fix to implement IP restrictions to on-campus
or campus VPN IP addresses.

It turned out that the change was not effective, and causes errors
in Nginx, so is being reverted by this commit, returning the
"main" branch to the same configuration as in the "1.0.0" version.

The previous commit and "1.0.1-rc1" tag are being kept for historical/archival
purposes.

https://issues.umd.edu/browse/LIBASPACE-298