LIBASPACE-310. Updated config to use umd-handle for handle minting
Updated the configuration to use the "umd-handle" application for
handle minting.

Updated "aspace-umd-lib-handle-service" plugin version in
"docker_config/archivesspace/config/plugins" to 2.0.0, which is the
version that provides the handle minting service using umd-handle.

In docker_config/archivesspace/archivesspace/config/config.rb,
added the environment variables needed to communicate with the
umd-handle service.

https://issues.umd.edu/browse/LIBASPACE-310