LIBASPACE-333. Updated "docker_config/archivespace/locales/enums/en.yml"
Starting in v3.1.0, the stock "en.yml" was sorted into alphanbetical
order by keys, which made the a "diff" comparison of the file to find
UMD customizations difficult.

To simplify finding the UMD customizations, copied the stock "en.yml"
file, and then updated with UMD customizations (comparing with the
v2.8.0 en.yml file).

https://issues.umd.edu/browse/LIBASPACE-333