LIBASPACE-333. Added local development environment setup instructions
Added local development environment setup instructions in
"docs/DevelopmentSetup.md". These instructions supersede those in the
https://confluence.umd.edu/display/LIB/ArchivesSpace+Development+Environment
Confluence page, which are now obsolete.

Added the development setup instructions to this repository, instead of
updating the Confluence page, as the setup instructions will likely
change as ArchivesSpace evolves, and this repository is most closely
tied to changes in the ArchivesSpace version.

Added link to the page from the README.md.

https://issues.umd.edu/browse/LIBASPACE-333