LIBASPACE-347. Replace addressable 2.8.4. with 2.8.0 in "plugins.sh"
The ArchivesSpace front-end is locked to v2.8.0 of the "addressable" gem
(see "archivesspace/frontend/Gemfile.lock"), but the "addressable" gem
specified by "aspace-oauth" is not pinned, and so changes over time.

Updated the "plugins.sh" script to replace the v2.8.4 addressable gem
version added by the "aspace-oauth" gem with v2.8.0 used by
ArchivesSpace.

https://umd-dit.atlassian.net/browse/LIBASPACE-347