LIBASPACE-366. Updated "_skipnav.html.erb", based on ArchivesSpace 3.5.1
The stock "_skipnav.html.erb" file has been updated since the initial
inclusion of this file in "aspace-custom".

Comparing the UMD-customized file against the initial "base", file the
only difference was the "<% if !defined?(@search) %>" check around the
first line.

Therefore, copied the stock ArchivesSpace 3.5.1 file as the base for
the updated file, and restored the UMD customizations, along with
customization markers.

https://umd-dit.atlassian.net/browse/LIBASPACE-366