LIBASPACE-367. Remove focus traversal customization - _skipnav.html.erb
In reviewing the "aspace-custom" customizations for the
ArchivesSpace 3.5.1 upgrade, it is unclear why this file, which modifies
the ArchivesSpace focus traversal order when using keyboard input
was actually needed (original issue where the file was added was
LIBASPACE-219).

After discussion with the stakeholders, it was decided that this
customization was not needed and so it was removed.

https://umd-dit.atlassian.net/browse/LIBASPACE-367