LIBASPACE-366. Update DevelopmentSetup.md with "aspace_yale_accessions"
Updated the DevelopmentSetup.md instructions to:

1) Add the "aspace_yale_accessions" plugin for use in the front-end
(staff) interface, as it has UMD customizations.

2) Re-arrange the steps so that the "run" instructions are after
both the stock ArchiveSpace and (optional) UMD customization setup
instructions. This is needed because the "aspace_yale_accessions"
plugin includes a database migration, so separating the "setup" and
"run" steps makes the overall flow the same regardless of whether
stock ArchivesSpace or UMD customizations are used.

https://umd-dit.atlassian.net/browse/LIBASPACE-366