LIBASPACE-366. Updated non-customized plugin versions
Updated the versions of plugins that UMD has not customized to the
latest tagged version in their respective repositories.

For hudmol/digitization_work_order, updated to v1.13. The latest is v2.0,
but it is not clear if that is intended for the current ArchivesSpace,
or the next major version of ArchivesSpace, (which will require changes
to plugins).

https://umd-dit.atlassian.net/browse/LIBASPACE-366