LIBASPACE-366. Update CAS config to support updated "aspace-oauth" plugin
To support the latest “lyrasis/aspace-oauth” version (3.5.0), needed to
update the “fetch_raw_info” parameter in “config.rb” to accept an
additional "rawxml" argument. This is due to the changes made in
https://github.com/dlindahl/omniauth-cas/commit/5fcb3890d637587e1da8c037e85d059ec6e21064

Converted to a “Proc” based on the example in the “omniauth-cas” README -
https://github.com/dlindahl/omniauth-cas/tree/v2.0.0, which may
future-proof us against future parameter changes.

https://umd-dit.atlassian.net/browse/LIBASPACE-366