LIBASPACE-379. Workaround for gem version error on startup
Modifies the "rubyzip" gem version in the
"digitization_work_order/Gemfile.lock" file to prevent the following
error at startup:

```
INFO: An exception happened during JRuby-Rack startup
Could not find rubyzip-2.3.2 in any of the sources
...
WARNING: ERROR: initialization failed
org.jruby.rack.RackInitializationException: Could not find rubyzip-2.3.2 in any of the sources
...
```

This seems to be a re-emergence of the “gem version” issue we saw in
ArchivesSpace 3.3.1, see

* https://umd-dit.atlassian.net/browse/LIBASPACE-333
* https://umd-dit.atlassian.net/browse/LIBASPACE-357

this time caused by the “hudmol/digitization_work_order” dependency,
which has a transitive dependency on an unconstrained “rubyzip” gem via
the its use of the “write_xlsx” gem.

Updated the "README.md" file.

https://umd-dit.atlassian.net/browse/LIBASPACE-379