LIBASPACE-389. Update Dockerfile base image to eclipse-temurin:8u462-b08
Updated the base Docker image used for the Java JRE used to
"eclipse-temurin:8u462-b08-jre", as the "openjdk:8u265-jre" was
using the Debian "buster" OS version, which has been archived.

Java v8u462-b08 is the latest at the current time, and the
"eclipse-temurin" image was selected because it has been used
successfully in DRUM/MD-SOAR and in fcrepo.

https://umd-dit.atlassian.net/browse/LIBASPACE-389