LIBASPACE-389. Docker image “ubuntu” user renamed to “aspace”
The “eclipse-temurin” base image already includes a “ubuntu” user with
UID/GID 1000, so also modified the “Dockerfile” to rename the “ubuntu”
user to “aspace”.

This is necessary because the “umd-lib/k8s-aspace” Kubernetes
configuration assumes that the “aspace” user has UID/GID 1000 when
setting file permissions.

While it is possible to simply add an additional “aspace” user with a
different UID/GID, such as 1001, and update the “umd-lib/k8s-aspace”
Kubernetes configuration, renaming the existing user seemed slightly
better to avoid having an extraneous user in the container.

https://umd-dit.atlassian.net/browse/LIBASPACE-389